Move preorder_traversal to sympy.core.basic
This will allow it to be used everywhere, and in particular in the core.
