+++ sympy/utilities/iterables.py
@@ -147,77 +147,6 @@ def postorder_traversal(node):
                 yield subtree
     yield node
 
-class preorder_traversal(object):
-    """
-    Do a pre-order traversal of a tree.
-
-    This iterator recursively yields nodes that it has visited in a pre-order
-    fashion. That is, it yields the current node then descends through the tree
-    breadth-first to yield all of a node's children's pre-order traversal.
-
-    Parameters
-    ----------
-    node : sympy expression
-        The expression to traverse.
-
-    Returns
-    -------
-    subtree : sympy expression
-        All of the subtrees in the tree.
-
-    Examples
-    --------
-    >>> from sympy import symbols
-    >>> from sympy.utilities.iterables import preorder_traversal
-    >>> from sympy.abc import x, y, z
-    >>> set(preorder_traversal((x+y)*z)) == set([z, x + y, z*(x + y), x, y])
-    True
-
-    """
-    def __init__(self, node):
-        self._skip_flag = False
-        self._pt = self._preorder_traversal(node)
-
-    def _preorder_traversal(self, node):
-        yield node
-        if self._skip_flag:
-            self._skip_flag = False
-            return
-        if isinstance(node, Basic):
-            for arg in node.args:
-                for subtree in self._preorder_traversal(arg):
-                    yield subtree
-        elif iterable(node):
-            for item in node:
-                for subtree in self._preorder_traversal(item):
-                    yield subtree
-
-    def skip(self):
-        """
-        Skip yielding current node's (last yielded node's) subtrees.
-
-        Examples
-        --------
-        >>> from sympy import symbols
-        >>> from sympy.utilities.iterables import preorder_traversal
-        >>> from sympy.abc import x, y, z
-        >>> pt = preorder_traversal((x+y*z)*z)
-        >>> for i in pt:
-        ...     print i
-        ...     if i == x+y*z:
-        ...             pt.skip()
-        z*(x + y*z)
-        z
-        x + y*z
-        """
-        self._skip_flag = True
-
-    def next(self):
-        return self._pt.next()
-
-    def __iter__(self):
-        return self
-
 def interactive_traversal(expr):
     """Traverse a tree asking a user which branch to choose. """
     from sympy.printing import pprint
